+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## v0.74.24, 1 November 2018
+
+- Yarn: Fix resolved url in lockfile for dependencies installed from GitHub using the
+  shorthand syntax (without host)
+
+## v0.74.23, 1 November 2018
+
+- Python: Handle upper bound changes when new version has fewer digits than old
+  one
+
 ## v0.74.22, 31 October 2018
 
 - Python: Handle custom python versions when using Poetry